requirements.txt
langchain
langchain_community
langchain_huggingface
faiss-cpu==1.7.4
pypdf
huggingface_hub
flask
python-dotenv
sentence-transformers